grocery_store/grocery/models.py
from django.contrib.auth.models import User
from django.db import models
from django.utils.text import slugify
from mptt.models import MPTTModel, TreeForeignKey


class Category(MPTTModel):
    name = models.CharField(max_length=255)
    slug = models.SlugField(unique=True)
    image = models.ImageField(upload_to='category_images/')
    parent = TreeForeignKey('self', on_delete=models.CASCADE, null=True, blank=True, related_name='categories')
    verbose_name = 'category'
    verbose_name_plural = 'categories'

    def save(self, *args, **kwargs):
        if not self.slug:
            self.slug = slugify(self.name)
        super().save(*args, **kwargs)

    class MPTTMeta:
        order_insertion_by = ['name']

    def __str__(self):
        return self.name


class Product(models.Model):
    name = models.CharField(max_length=255)
    slug = models.SlugField(unique=True)
    price = models.DecimalField(max_digits=10, decimal_places=2)
    category = TreeForeignKey(
        Category, related_name="products", on_delete=models.CASCADE
    )
    image_large = models.ImageField(upload_to='product_images/large/')
    image_medium = models.ImageField(upload_to='product_images/medium/')
    image_small = models.ImageField(upload_to='product_images/small/')

    def save(self, *args, **kwargs):
        if not self.slug:
            self.slug = slugify(self.name)
        super().save(*args, **kwargs)

    def __str__(self):
        return self.name


class CartItem(models.Model):
    user = models.ForeignKey(User, on_delete=models.CASCADE)
    product = models.ForeignKey(Product, on_delete=models.CASCADE)
    quantity = models.PositiveIntegerField(default=1)
